+++ core/edge_mining/__main__.py
@@ -73,6 +73,8 @@ async def main_async():
             optimization_service=services.optimization_service,
             logger=logger,
             settings=settings,
+            system_configuration=services.configuration_service.get_system_configuration(),
+            event_bus=services.event_bus,
             home_load_history_service=services.home_load_history_service,
             load_forecast_training_service=services.load_forecast_training_service,
         )